requirements.txt
google-generativeai
questionary
requests